% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ipynb_to_blogdown.R
\name{ipynb_to_blogdown_post}
\alias{ipynb_to_blogdown_post}
\title{Convert Jupyter Notebook to Blogdown Post}
\usage{
ipynb_to_blogdown_post(
  ipynb_file,
  chunk_options = "",
  title = "",
  author = "",
  date = today(),
  ...
)
}
\arguments{
\item{ipynb_file}{path to the jupyter notebook}

\item{chunk_options}{string. Passes options to kitr::opts_chunk$set(). Defaults to ""}

\item{title}{title of the markdown. Defaults to ""}

\item{author}{author. Defaults to ""}

\item{date}{day of creation as string. Defaults to today()}

\item{...}{arguments passed to \code{\linl{md_to_Rmd}} such as the python path used for reticulate}
}
\description{
Converts a jupyter notebook to a Blogdown post.
}
