% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ipynb_to_blogdown.R
\name{ipynb_to_md}
\alias{ipynb_to_md}
\title{Jupyter Notebook to markdown}
\usage{
ipynb_to_md(ipynb_file)
}
\arguments{
\item{ipynb_file}{path to the jupyter notebook}
}
\description{
Uses nbconvert and nbstripout to convert the jupyter notebook to
a clean markdown.
}
