% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ipynb_to_blogdown.R
\name{ipynb_to_Rmd}
\alias{ipynb_to_Rmd}
\title{Convert Jupyter Notebook to Rmd}
\usage{
ipynb_to_Rmd(ipynb_file, output_file = "", blogdown, ...)
}
\arguments{
\item{ipynb_file}{path to the jupyter notebook}

\item{output_file}{path to the generated Rmd-file. Defaults to the same name as the input file}

\item{blogdown}{logical indicating if a blogdown header should be generated. Otherwise a normal .Rmd header is generated.}

\item{...}{arguments to be passed to \code{\link{md_to_Rmd}} such as title, author, or date}
}
\description{
Converts a jupyter notebook to rmd.
}
