% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ipynb_to_blogdown.R
\name{ipynb_to_Rmd_text}
\alias{ipynb_to_Rmd_text}
\title{Convert Jupyter Notebook to Rmd}
\usage{
ipynb_to_Rmd_text(
  ipynb_file,
  chunk_options = "",
  python_path = reticulate::py_config()$python
)
}
\arguments{
\item{ipynb_file}{path to the jupyter notebook}

\item{chunk_options}{string. Passes options to \code{\link[knitr:opts_chunk]{knitr::opts_chunk$set()}}. Defaults to ""}

\item{python_path}{python path as string for reticulate. Defaults to the
one detected by \code{\link[reticulate:py_config]{reticulate::py_config()}}}
}
\description{
Converts a jupyter notebook to Rmd style md. Returns the string without YAML header.
}
