% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ipynb_to_blogdown.R
\name{make_slug}
\alias{make_slug}
\title{Make a Slug out of a Title}
\usage{
make_slug(title)
}
\arguments{
\item{title}{title as string}
}
\description{
This function parses a title into a slug
}
\examples{
make_slub("This is a nice title")
}
