% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ipynb_to_blogdown.R
\name{md_to_Rmd}
\alias{md_to_Rmd}
\title{Convert a md to rmd}
\usage{
md_to_Rmd(
  md_file,
  output_file = "",
  chunk_options = "",
  title = "",
  author = "",
  date = lubridate::today(),
  blogdown = FALSE,
  python_path = reticulate::py_config()$python
)
}
\arguments{
\item{md_file}{path to the markdown file}

\item{output_file}{path to the generated Rmd-file. Defaults to the same name as the input file}

\item{chunk_options}{string. Passes options to kitr::opts_chunk$set(). Defaults to ""}

\item{title}{title of the markdown. Defaults to ""}

\item{author}{author. Defaults to ""}

\item{date}{day of creation as string. Defaults to \code{\link{lubridate::today()}}}

\item{blogdown}{logical indicating if a blogdown header should be generated. Otherwise a normal .Rmd header is generated. Defaults to FALSE}

\item{python_path}{python path as string for reticulate. Defaults to the one detected by \code{\link{reticulate::py_config()}}}
}
\description{
Converts a markdown file containing python code chunks to a
Rmd file using reticulate
}
