% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ipynb_to_blogdown.R
\name{md_to_Rmd_text}
\alias{md_to_Rmd_text}
\title{Convert a md to rmd}
\usage{
md_to_Rmd_text(
  md_file,
  chunk_options = "",
  python_path = reticulate::py_config()$python
)
}
\arguments{
\item{md_file}{path to the markdown file}

\item{chunk_options}{string. Passes options to \code{\link[knitr:opts_chunk]{knitr::opts_chunk$set()}}. Defaults to ""}

\item{python_path}{python path as string for reticulate. Defaults to the one detected by \code{\link[reticulate:py_config]{reticulate::py_config()}}}
}
\description{
Converts a markdown file containing python code chunks to a
Rmd file with a reticulate chunk. Returns the Rmd string without YAML header
}
