% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ipynb_to_blogdown.R
\name{new_post_from_ipynb}
\alias{new_post_from_ipynb}
\title{Creates a new Blogdown Post from a Jupyter Notebook}
\usage{
new_post_from_ipynb(
  ipynb_file,
  chunk_options = "",
  python_path = reticulate::py_config()$python,
  title = "",
  kind = "",
  open = interactive(),
  author = getOption("blogdown.author"),
  categories = NULL,
  tags = NULL,
  date = Sys.Date(),
  file = NULL,
  slug = NULL,
  title_case = getOption("blogdown.title_case"),
  subdir = getOption("blogdown.subdir", "post"),
  ext = getOption("blogdown.ext", ".Rmd")
)
}
\arguments{
\item{ipynb_file}{path to the jupyter notebook}

\item{chunk_options}{global chunk options as charachter vector. Passes options to \code{\link[knitr:opts_chunk]{knitr::opts_chunk$set()}}. Defaults to ""}

\item{python_path}{python path as string for reticulate. Defaults to the one detected by \code{\link[reticulate:py_config]{reticulate::py_config()}}}

\item{title}{title of the markdown. Defaults to the name of the notebook}

\item{kind}{content type to create}

\item{open}{whether to open the generated file after creation}

\item{author}{author.}

\item{categories}{character vector of category names}

\item{tags}{character vector of tag names}

\item{date}{date of the post. Defaults to \code{Sys.Date()}}

\item{file}{filename of the post. Will be automatically generated by default from the title}

\item{slug}{slug of the post. Also automatically generated by default from the title}

\item{title_case}{function to convert the title to title case}

\item{subdir}{if specified the post will be generated in a subdirectoy of 'content/'}

\item{ext}{extension of the generated file. Defaults to .Rmd}
}
\value{
Returns the file name of the newly generated file.
}
\description{
Creates a new blogdown post and appends the jupyter
notebook content as Rmd-style text. The notebook
gets transformed into a runnable Rmd with python chunks.
}
\seealso{
\code{\link[blogdown:new_post]{blogdown::new_post()}} for which this function is a wrapper.
}
