% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ipynb_to_blogdown.R
\name{remove_pngs}
\alias{remove_pngs}
\title{Removes Plot Outputs}
\usage{
remove_pngs(md_text)
}
\arguments{
\item{md_text}{markdown text as string}
}
\description{
This function removes any png plot outputs.
}
