% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ipynb_to_blogdown.R
\name{Rmd_from_ipynb}
\alias{Rmd_from_ipynb}
\title{Creates an Rmd file from a Jupyter Notebook}
\usage{
Rmd_from_ipynb(
  ipynb_file,
  chunk_options = "",
  python_path = reticulate::py_config()$python,
  open = interactive(),
  title = "",
  author = "",
  date = Sys.Date()
)
}
\arguments{
\item{ipynb_file}{path to the jupyter notebook}

\item{chunk_options}{global chunk options as charachter vector. Passes options to kitr::opts_chunk$set(). Defaults to ""}

\item{python_path}{python path as string for reticulate. Defaults to the one detected by \code{\link{reticulate::py_config()}}}

\item{open}{whether to open the generated file after creation}

\item{title}{title of the markdown as string. Defaults to ""}

\item{author}{author as string. Defaults to ""}

\item{date}{day of creation as string. Defaults to \code{Sys.Date()}}
}
\description{
Creates an Rmd file from a Jupyter Notebook
}
