% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ipynb_to_blogdown.R
\name{transform_python_cells}
\alias{transform_python_cells}
\title{Transform Python Markdown Code Snippets}
\usage{
transform_python_cells(md_text)
}
\arguments{
\item{md_text}{markdown text as string}
}
\description{
This function takes a markdown-text as string-input and transforms
each python code snippet in an executable python cell for Rmarkdown
}
