% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ipynb_to_blogdown.R
\name{add_yaml}
\alias{add_yaml}
\title{Add a YAML header}
\usage{
add_yaml(md_text, title = "", author = "", date = today(), blogdown = FALSE)
}
\arguments{
\item{md_text}{markdown text as string}

\item{title}{title of the markdown as string. Defaults to ""}

\item{author}{author as string. Defaults to ""}

\item{date}{day of creation as string. Defaults to \code{\link{today()}}}

\item{blogdown}{logical indicating if a blogdown header should be generated. Otherwise a normal .Rmd header is generated. Defaults to FALSE}
}
\description{
Adds a YAML header to your markdown
}
